fix(artwork): cast reused parameter in accounting publish statement
Postgres deduces one type per placeholder; $7 appeared as the bigint
failure_count assignment and in an untyped CASE comparison, failing the
snapshot publish with SQLSTATE 42P08 on every live refresh. Cast the
comparison explicitly. Audited every other reused placeholder in the
artwork SQL for the same pattern; all are type-consistent.

+++ internal/metadata/artwork_storage.go
@@ -202,7 +202,7 @@ const artworkAccountingPublishSQL = `
 		coverage_limit_reason = $9,
 		adoption_index_bytes = $10,
 		adoption_index_objects = $11,
-		last_error = CASE WHEN $7 > 0 THEN 'inventory refresh had failures' ELSE '' END,
+		last_error = CASE WHEN $7::bigint > 0 THEN 'inventory refresh had failures' ELSE '' END,
 		updated_at = NOW()
 	WHERE singleton`
